Extract MCP specific logic into it's own class/module
