+++ chatlas/_chat.py
@@ -1,18 +1,14 @@
 from __future__ import annotations
 
-import asyncio
 import copy
 import inspect
 import os
 import sys
 import traceback
 import warnings
-from contextlib import AsyncExitStack
-from dataclasses import dataclass, field
 from pathlib import Path
 from threading import Thread
 from typing import (
-    TYPE_CHECKING,
     Any,
     AsyncGenerator,
     AsyncIterator,
@@ -46,6 +42,7 @@ from ._display import (
     MockMarkdownDisplay,
 )
 from ._logging import log_tool_error
+from ._mcp_manager import MCPManager
 from ._provider import Provider
 from ._tools import Tool, ToolRejectError
 from ._turn import Turn, user_turn
@@ -66,44 +63,6 @@ method of a [](`~chatlas.Chat`) instance.
 CompletionT = TypeVar("CompletionT")
 
 EchoOptions = Literal["output", "all", "none", "text"]
-
-if TYPE_CHECKING:
-    from mcp import ClientSession as MCPClientSession
-
-
-@dataclass
-class MCPSessionInfo:
-    name: str
-    task: asyncio.Task | None = None
-    ready_event: asyncio.Event = field(default_factory=asyncio.Event)
-    shutdown_event: asyncio.Event = field(default_factory=asyncio.Event)
-    exit_stack: AsyncExitStack = field(default_factory=AsyncExitStack)
-    error: asyncio.CancelledError | Exception | None = None
-
-
-@dataclass
-class HTTPMCPSessionInfo(MCPSessionInfo):
-    url: str = ""
-    transport_kwargs: dict[str, Any] = field(default_factory=dict)
-
-
-@dataclass
-class STDIOMCPSessionInfo(MCPSessionInfo):
-    command: str = ""
-    args: list[str] = field(default_factory=list)
-    transport_kwargs: dict[str, Any] = field(default_factory=dict)
-
-
-def try_import_mcp():
-    try:
-        import mcp
-
-        return mcp
-    except ImportError:
-        raise ImportError(
-            "The `mcp` package is required to connect to MCP servers. "
-            "Install it with `pip install mcp`."
-        )
 
 
 class Chat(Generic[SubmitInputArgsT, CompletionT]):
@@ -147,7 +106,7 @@ class Chat(Generic[SubmitInputArgsT, CompletionT]):
             "rich_console": {},
             "css_styles": {},
         }
-        self._mcp_sessions: dict[str, MCPSessionInfo] = {}
+        self._mcp_manager = MCPManager()
 
     def get_turns(
         self,
@@ -958,8 +917,8 @@ class Chat(Generic[SubmitInputArgsT, CompletionT]):
         *,
         name: str,
         url: str,
-        include_tools: Optional[list[str]] = None,
-        exclude_tools: Optional[list[str]] = None,
+        include_tools: Sequence[str] = (),
+        exclude_tools: Sequence[str] = (),
         namespace: Optional[str] = None,
         transport_kwargs: Optional[dict[str, Any]] = None,
     ):
@@ -1056,36 +1015,29 @@ class Chat(Generic[SubmitInputArgsT, CompletionT]):
         )
         ```
         """
+        if isinstance(exclude_tools, str):
+            exclude_tools = [exclude_tools]
+        if isinstance(include_tools, str):
+            include_tools = [include_tools]
 
-        if name in self._mcp_sessions:
-            raise ValueError(f"MCP Session {name} already exists.")
-
-        _ = try_import_mcp()
-
-        session_info = HTTPMCPSessionInfo(
+        session_info = await self._mcp_manager.register_http_stream_tools(
             name=name,
             url=url,
+            include_tools=include_tools,
+            exclude_tools=exclude_tools,
+            namespace=namespace,
             transport_kwargs=transport_kwargs or {},
         )
 
-        # Launch background task that runs until MCP session is *shutdown*
-        # N.B. this is needed since mcp session can't be opened in one task and closed in another
-        asyncio.create_task(
-            self._init_mcp_session(
-                session_info=session_info,
-                include_tools=include_tools,
-                exclude_tools=exclude_tools,
-                namespace=namespace,
+        overlapping_tools = set(self._tools.keys()) & set(session_info.tools)
+        if overlapping_tools:
+            await self._mcp_manager.close_sessions([name])
+            raise ValueError(
+                f"The following tools are already registered: {overlapping_tools}. "
+                "Consider providing a namespace when registering this MCP server "
+                "to avoid name collisions."
             )
-        )
-
-        # Wait for a ready event from the background task (signals that tools are registered)
-        await session_info.ready_event.wait()
-
-        if session_info.error:
-            raise RuntimeError(
-                f"Failed to register tools from MCP server '{name}' at URL '{url}'"
-            ) from session_info.error
+        self._tools.update(session_info.tools)
 
     async def register_mcp_tools_stdio_async(
         self,
@@ -1093,8 +1045,8 @@ class Chat(Generic[SubmitInputArgsT, CompletionT]):
         name: str,
         command: str,
         args: list[str],
-        include_tools: Optional[list[str]] = None,
-        exclude_tools: Optional[list[str]] = None,
+        include_tools: Sequence[str] = (),
+        exclude_tools: Sequence[str] = (),
         namespace: Optional[str] = None,
         transport_kwargs: Optional[dict[str, Any]] = None,
     ):
@@ -1191,38 +1143,31 @@ class Chat(Generic[SubmitInputArgsT, CompletionT]):
         )
         ```
         """
+        if isinstance(exclude_tools, str):
+            exclude_tools = [exclude_tools]
+        if isinstance(include_tools, str):
+            include_tools = [include_tools]
 
-        if name in self._mcp_sessions:
-            raise ValueError(f"MCP Session {name} already exists.")
-
-        _ = try_import_mcp()
-
-        # Launch background task that runs until MCP session is *shutdown*
-        session_info = STDIOMCPSessionInfo(
+        session_info = await self._mcp_manager.register_stdio_tools(
             name=name,
             command=command,
             args=args,
+            include_tools=include_tools,
+            exclude_tools=exclude_tools,
+            namespace=namespace,
             transport_kwargs=transport_kwargs or {},
         )
 
-        # Launch a background task to initialize the MCP server
-        # N.B. this is needed since mcp session can't be opened in one task and closed in another
-        asyncio.create_task(
-            self._init_mcp_session(
-                session_info=session_info,
-                include_tools=include_tools,
-                exclude_tools=exclude_tools,
-                namespace=namespace,
+        overlapping_tools = set(self._tools.keys()) & set(session_info.tools)
+        if overlapping_tools:
+            await self._mcp_manager.close_sessions([name])
+            raise ValueError(
+                f"The following tools are already registered: {overlapping_tools}. "
+                "Consider providing a namespace when registering this MCP server "
+                "to avoid name collisions."
             )
-        )
 
-        # Wait for a ready event from the background task (signals that tools are registered)
-        await session_info.ready_event.wait()
-
-        if session_info.error:
-            raise RuntimeError(
-                f"Failed to register tools from MCP server '{name}' with command '{command} {args}'"
-            ) from session_info.error
+        self._tools.update(session_info.tools)
 
     async def cleanup_mcp_tools(self, names: Optional[Sequence[str]] = None):
         """
@@ -1235,187 +1180,22 @@ class Chat(Generic[SubmitInputArgsT, CompletionT]):
 
         Parameters
         ----------
-        name
+        names
             If provided, only clean up the tools and session associated
-            with this name. If not provided, clean up all registered MCP tools and sessions.
+            with these names. If not provided, clean up all registered MCP tools and sessions.
 
         Returns
         -------
         None
         """
-        if names is None:
-            names = list(self._mcp_sessions.keys())
+        # MCPManager will return the names of tools that were removed
+        closed_sessions = await self._mcp_manager.close_sessions(names)
 
-        if isinstance(names, str):
-            names = [names]
-
-        for x in names:
-            if x not in self._mcp_sessions:
-                warnings.warn(
-                    f"No MCP session found with name '{x}'. Skipping cleanup.",
-                    stacklevel=2,
-                )
-                continue
-            session = self._mcp_sessions[x]
-            session.shutdown_event.set()
-            if session.task is not None:
-                await session.task
-
-    async def _init_mcp_session(
-        self,
-        session_info: "MCPSessionInfo",
-        include_tools: Optional[list[str]] = None,
-        exclude_tools: Optional[list[str]] = None,
-        namespace: Optional[str] = None,
-    ):
-        session_info.task = asyncio.current_task()
-        self._mcp_sessions[session_info.name] = session_info
-
-        try:
-            # Establish the MCP session
-            if isinstance(session_info, HTTPMCPSessionInfo):
-                session = await self._init_mcp_http_session(session_info)
-            elif isinstance(session_info, STDIOMCPSessionInfo):
-                session = await self._init_mcp_stdio_session(session_info)
-            else:
-                raise TypeError(
-                    f"Unsupported session type: {type(session_info).__name__}. "
-                    "Expected HTTPMCPSessionInfo or STDIOMCPSessionInfo."
-                )
-
-            # Request the available MCP tools, filter/namespace them,
-            # and convert into to our Tool class
-            tools = await self._request_mcp_tools(
-                session=session,
-                include_tools=include_tools,
-                exclude_tools=exclude_tools,
-                namespace=namespace,
-            )
-
-            # Register tools with the chat (assuming no name collisions)
-            overlapping_tools = set(self._tools.keys()) & set(tools.keys())
-            if overlapping_tools:
-                raise ValueError(
-                    f"The following tools are already registered: {overlapping_tools}. "
-                    "Consider providing a namespace when registering this MCP server "
-                    "to avoid name collisions."
-                )
-            self._tools.update(tools)
-
-        except (asyncio.CancelledError, Exception) as e:
-            # Remember error so we can handle in the main task
-            session_info.error = e
-            # And also close the exit stack in case the connection was opened
-            try:
-                await session_info.exit_stack.aclose()
-            except Exception:
-                pass
-            return
-        finally:
-            # Whether successful or not, set ready state to prevent deadlock
-            session_info.ready_event.set()
-
-        # If successful, wait for shutdown signal
-        await session_info.shutdown_event.wait()
-
-        # On shutdown close connection to MCP server
-        # This is why we're using a background task in the 1st place...
-        # we must close in the same task that opened the session
-        await session_info.exit_stack.aclose()
-
-        # Also delete the session and corresponding tools
-        del self._mcp_sessions[session_info.name]
-        for tool_name in tools.keys():
-            del self._tools[tool_name]
-
-    async def _init_mcp_http_session(self, session_info: "HTTPMCPSessionInfo"):
-        from mcp import ClientSession
-        from mcp.client.streamable_http import streamablehttp_client
-
-        read, write, _ = await session_info.exit_stack.enter_async_context(
-            streamablehttp_client(
-                session_info.url,
-                **session_info.transport_kwargs,
-            )
-        )
-        session = await session_info.exit_stack.enter_async_context(
-            ClientSession(read, write)
-        )
-        await session.initialize()
-        return session
-
-    async def _init_mcp_stdio_session(self, session_info: "STDIOMCPSessionInfo"):
-        # Try to initialize the MCP session (and cleanup if it fails)
-        mcp = try_import_mcp()
-        from mcp.client.stdio import stdio_client
-
-        command = session_info.command
-        args = session_info.args
-
-        server_params = mcp.StdioServerParameters(
-            command=command,
-            args=args,
-            **session_info.transport_kwargs,
-        )
-
-        transport = await session_info.exit_stack.enter_async_context(
-            stdio_client(server_params)
-        )
-        session = await session_info.exit_stack.enter_async_context(
-            mcp.ClientSession(*transport)
-        )
-        await session.initialize()
-        return session
-
-    async def _request_mcp_tools(
-        self,
-        session: "MCPClientSession",
-        include_tools: list[str] | None = None,
-        exclude_tools: list[str] | None = None,
-        namespace: str | None = None,
-    ):
-        if include_tools and exclude_tools:
-            raise ValueError("Cannot specify both include_tools and exclude_tools.")
-
-        # Request the MCP tools available
-        response = await session.list_tools()
-        mcp_tools = response.tools
-        tool_names = set(x.name for x in mcp_tools)
-
-        # Warn if tools are mis-specified
-        include = set(include_tools or [])
-        missing_include = include.difference(tool_names)
-        if missing_include:
-            warnings.warn(
-                f"Specified include_tools {missing_include} did not match any tools from the MCP server. "
-                f"The tools available are: {tool_names}",
-                stacklevel=2,
-            )
-        exclude = set(exclude_tools or [])
-        missing_exclude = exclude.difference(tool_names)
-        if missing_exclude:
-            warnings.warn(
-                f"Specified exclude_tools {missing_exclude} did not match any tools from the MCP server. "
-                f"The tools available are: {tool_names}",
-                stacklevel=2,
-            )
-
-        # Filter the tool names
-        if include:
-            tool_names = include.intersection(tool_names)
-        if exclude:
-            tool_names = tool_names.difference(exclude)
-
-        # Apply namespace and convert to chatlas.Tool instances
-        res: dict[str, Tool] = {}
-        for tool in mcp_tools:
-            if tool.name not in tool_names:
-                continue
-            if namespace:
-                tool.name = f"{namespace}.{tool.name}"
-            res[tool.name] = Tool.from_mcp(session=session, mcp_tool=tool)
-
-        return res
+        # Remove those tools from the main tools dict
+        for session in closed_sessions:
+            for tool_name in session.tools:
+                if tool_name in self._tools:
+                    del self._tools[tool_name]
 
     def register_tool(
         self,
